+++ client/app/(main)/dashboard/page.tsx
@@ -1,0 +1,5 @@
+function Dashboard(): JSX.Element {
+  return <div className="flex min-h-screen flex-col items-center space-y-20">Dashboard page</div>;
+}
+
+export default Dashboard;
