+++ client/app/(main)/page.tsx
@@ -1,26 +1,16 @@
-'use client';
+'use client'
 
-import Image from 'next/image';
-import { useState, useEffect } from 'react';
+import { useState } from 'react';
 
-import Link from 'next/link';
-import Recommendations from './_components/recommendations';
-import LoadingSpinner from './_components/loadingSpinner';
+import Recommendations from '../_components/recommendations';
+import LoadingSpinner from '../_components/loadingSpinner';
 
 export default function Home(): JSX.Element {
   const [recommendedGifts, setRecommendedGifts] = useState<string[]>([]);
   const [giftsImages, setGiftImages] = useState<string[]>([]);
   const [loading, setLoading] = useState<boolean>(false);
-  const [userData, setUserData] = useState<string>('');
 
-  useEffect(() => {
-    const storedUserData = localStorage.getItem('userData');
-    if (storedUserData) {
-      const parsedUserData = JSON.parse(storedUserData);
-      setUserData(parsedUserData);
-    }
-  }, []);
-
+  
   // form submission, submitting user's inputs and fetching gift recommendations from the server
   const handleSubmit = async (
     event: React.FormEvent<HTMLFormElement>
@@ -49,49 +39,7 @@ export default function Home(): JSX.Element {
     setLoading(false);
   };
   return (
-    <main className="flex min-h-screen flex-col items-center p-24 space-y-40">
-      <div className="z-10 max-w-5xl w-full items-center justify-between font-mono text-sm lg:flex">
-        <div className="fixed bottom-0 left-0 flex h-48 w-full items-end justify-center bg-gradient-to-t from-white via-white dark:from-black dark:via-black lg:static lg:h-auto lg:w-auto lg:bg-none space-y-40">
-          <a
-            className="pointer-events-none flex place-items-center gap-2 p-8 lg:pointer-events-auto lg:p-0"
-            href=""
-            target="_blank"
-            rel="noopener noreferrer"
-          >
-            <Image
-              src="/logo.svg"
-              alt="giftsAI logo"
-              width={100}
-              height={24}
-              priority
-            />
-          </a>
-        </div>
-        <div className="flex gap-2">
-          <Link
-            className="fixed left-0 top-0 flex w-full justify-center border-b border-gray-300 pb-6 pt-8 backdrop-blur-2xl dark:border-neutral-800 dark:bg-sky-300 dark:text-white lg:static lg:w-auto lg:rounded-xl lg:border lg:bg-gray-200 lg:p-4 lg:dark:bg-transparent hover:bg-white hover:bg-opacity-10"
-            href="/create-account"
-          >
-            Create Account
-          </Link>
-          {userData ? (
-            <Link
-              className="fixed left-0 top-0 flex w-full justify-center border-b border-gray-300 pb-6 pt-8 backdrop-blur-2xl dark:border-neutral-800 dark:bg-sky-300 dark:text-black lg:static lg:w-auto lg:rounded-xl lg:border lg:bg-gray-200 lg:p-4 lg:dark:bg-sky-300 hover:bg-sky-200"
-              href="/"
-            >
-              Log Out
-            </Link>
-          ) : (
-            <Link
-              className="fixed left-0 top-0 flex w-full justify-center border-b border-gray-300 pb-6 pt-8 backdrop-blur-2xl dark:border-neutral-800 dark:bg-sky-300 dark:text-black lg:static lg:w-auto lg:rounded-xl lg:border lg:bg-gray-200 lg:p-4 lg:dark:bg-sky-300 hover:bg-sky-200"
-              href="/log-in"
-            >
-              Log In
-            </Link>
-          )}
-        </div>
-      </div>
-
+    <div className="flex min-h-screen flex-col items-center space-y-20">
       <form onSubmit={handleSubmit} className="flex flex-col gap-8">
         <div className="mb-32 grid text-center lg:max-w-5xl lg:w-full lg:mb-0 lg:grid-cols-4 lg:text-left">
           <div className="group rounded-lg border border-transparent px-5 py-4 transition-colors">
@@ -138,11 +86,10 @@ export default function Home(): JSX.Element {
         <LoadingSpinner />
       ) : (
         <Recommendations
-          userData={userData}
           recommendedGifts={recommendedGifts}
           giftImages={giftsImages}
         />
       )}
-    </main>
+    </div>
   );
 }
