--- client/app/dashboard/page.tsx
@@ -1,5 +1,0 @@
-function Dashboard(): JSX.Element {
-  return <>Dashboard page</>;
-}
-
-export default Dashboard;
